Extract a test env class
